finish aesthetics; refactor onResize() to be fully declarative
