+++ src/callbacks/onResize/annotateGrouping/vertically.js
@@ -1,6 +1,5 @@
 export default function vertically() {
     this.groupings.forEach(d => {
-        console.log(d);
         if (d.IDs.length) {
             const nIDs = d.IDs.length,
                 firstID = d.IDs[nIDs - 1],
@@ -16,8 +15,8 @@ export default function vertically() {
                     .attr({
                         x1: this.plot_width,
                         x2: this.plot_width + this.margin.right / 8,
-                        y1: y1 + this.y.rangeBand() / 4,
-                        y2: y1 + this.y.rangeBand() / 4
+                        y1: y1 + 3 * this.y.rangeBand() / 4,
+                        y2: y1 + 3 * this.y.rangeBand() / 4
                     }),
                 span = g
                     .append('line')
@@ -25,8 +24,8 @@ export default function vertically() {
                     .attr({
                         x1: this.plot_width + this.margin.right / 8,
                         x2: this.plot_width + this.margin.right / 8,
-                        y1: y1 + this.y.rangeBand() / 4,
-                        y2: y2 + 3 * this.y.rangeBand() / 4
+                        y1: y1 + 3 * this.y.rangeBand() / 4,
+                        y2: y2 + this.y.rangeBand()
                     }),
                 bottomBoundary = g
                     .append('line')
@@ -34,17 +33,17 @@ export default function vertically() {
                     .attr({
                         x1: this.plot_width,
                         x2: this.plot_width + this.margin.right / 8,
-                        y1: y2 + 3 * this.y.rangeBand() / 4,
-                        y2: y2 + 3 * this.y.rangeBand() / 4
+                        y1: y2 + this.y.rangeBand(),
+                        y2: y2 + this.y.rangeBand()
                     }),
                 annotation = g
                     .append('text')
                     .classed('annotation', true)
                     .attr({
                         x: this.plot_width,
-                        dx: 5 * this.margin.right / 8,
+                        dx: 4 * this.margin.right / 8,
                         y: y1,
-                        dy: this.y.rangeBand() / 2
+                        dy: this.y.rangeBand()
                     })
                     .text(
                         `${this.config.groupings.filter(
